I’ve finished developing the reporting functionality, everything works stably

+++ report-service/src/main/java/com/another/reportservice/entity/reportEntity/PerformerEfficiencyReport.java
@@ -13,5 +13,5 @@ public class PerformerEfficiencyReport {
     private Long successfulImplementation;
     private Long hiredTask;
     private Double AVGCompletionTime;
-    private Double userRating;
+    private Integer userRating;
 }
